requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
pytz==2023.3.post1
email-validator==2.1.0.post1
